fix(tests): Fix validation test compilation errors (Agent 4)
- Fixed ReferenceValidatorTest addError() method calls to use correct 3-parameter signature
- Fixed HistogramFindOptionsTest to use MatchAdjustmentOptions.builder() static method
- Part of TEST_COMPILATION_FIX_PLAN Agent 4 responsibilities

+++ library/src/main/java/io/github/jspinak/brobot/tools/actionhistory/ActionHistoryPersistence.java
@@ -202,7 +202,7 @@ public class ActionHistoryPersistence {
             .setMatchList(result.getMatchList())
             .setText(textResult)
             .setActionSuccess(result.isSuccess())
-            .setDuration((long) result.getDuration().toMillis())
+            .setDuration(result.getDuration().toMillis())
             .build();
         
         // Add to history
